Separated the summations into separate files so that tests can be written which isolate them more easily.
