+++ include/s1.h
@@ -1,0 +1,15 @@
+#ifndef S1_H
+#define S1_H
+#include "Environment.h"
+#include <iostream>
+#include <cmath>
+#include <iomanip>
+#include <primesieve.hpp>
+#include <primecount.hpp>
+#include <climits>
+#include "mpreal.h"
+#include "Primetools.h"
+
+mpfr::mpreal S1(uint64_t, mpfr::mpreal);
+
+#endif
