+++ include/s2.h
@@ -1,0 +1,21 @@
+#ifndef S2_H
+#define S2_H
+#include "Environment.h"
+#include "B2.h"
+#include "mobius.h"
+#include <iostream>
+#include <cmath>
+#include <iomanip>
+#include <primesieve.hpp>
+#include <primecount.hpp>
+#include <climits>
+#include "mpreal.h"
+
+long double T(long double);
+mpfr::mpreal S2(uint64_t, mpfr::mpreal);
+
+#endif
+
+static B2Class B2;
+static const mpfr::mpreal PI = 3.141592653589793238462643383279L;
+static const int J = 10;
