+++ src/mobius.cc
@@ -1,0 +1,50 @@
+#include "mobius.h"
+
+const uint64_t mobiusCutoff = 1000001;
+long long *mobiusTable;
+
+void mobius_setup() {
+  std::cout<<"seting up mobius table...";
+  mobiusTable = new long long[mobiusCutoff];
+  for (uint i = 0; i < mobiusCutoff; ++i) {
+    // mobiusTable[i] = -2000;
+    mobiusTable[i] = mobius_work(i);
+  }
+  std::cout<<"Done"<<std::endl;
+}
+
+long long mobius (long long x) {
+  if (x < mobiusCutoff) {
+    if (mobiusTable[x] == -2000)
+      mobiusTable[x] = mobius_work(x);
+    return mobiusTable[x];
+  }
+  return mobius_work(x);
+}
+
+long long mobius_work(long long x) {
+
+  long long k = 0;
+  if(x == 1)
+    return 1;
+  if(x == 2)
+    return -1;
+  if(x % 2 == 0) {
+    x = x / 2;
+    k++;
+    if(x % 2 == 0)
+      return 0;
+  }
+
+  for(long long i = 3; i*i <= x; i += 2) {
+    if(x % i == 0) {
+      x = x / i;
+      k++;
+      if(x % i == 0)
+        return 0;
+    }
+  }
+  return (k % 2 == 0) ? -1 : 1;
+}
+
+
